Move code for reading process parameters into input/process/parameter.rs
